Test the Hit Die directive
To implement rules that can be used to understand how class features
interact with character sheets, we'll need to be able to explain them
in an easily parseable manner.

The rules should be in Markdown for ease of use, editing, and portability.
So we'll create a small language that can be embedded in Markdown, which is
itself still perfectly valid Markdown, but can be parsed out and used.